fix dev mode not reloading code changes
hacky fix due to the fact that we load the app config at a time later
then when the namespace is first loaded, so our old wrap-env-middleware
cannot work as-is. and then immutant's own run function alternative that
can be used to run the server in dev mode has some super annoying
behaviour that can't be turned off (opening a browser on startup... who
the *fuck* actually likes this behaviour? don't you people keep a
browser open all the time when you're working on web dev anyway?)