better handling of user login/logout and passing user-id's to views
when logging in/out, there is a small window where the server's session
state and the client-side state and view subscriptions will not exactly
align and some view subscription updates (because
mtgcoll.client.auth/user-profile was updated) will get treated as
unauthorized.

this change is admittedly somewhat quick/hacky, but basically the idea
is to track the sente connection state (which is true when sente is both
connected and a handshake event has been received) in a reagent atom.
if false, we hide the entire ui and show a "connecting ..." message.

not super pretty, but it's only shown to the user briefly and it solves
the problem. i'll need to revisit this at some point and wrap it up
in some library code or something somewhere to make the whole method
prettier to use in the future